refactor more types in single place to export
