+++ src/fetch.ts
@@ -1,14 +1,6 @@
 import type { Readable } from "node:stream";
 import destr from "destr";
 import { withBase, withQuery } from "ufo";
-import type {
-  Fetch,
-  RequestInfo,
-  RequestInit,
-  Response,
-  ResponseType,
-  MappedType,
-} from "./types";
 import { createFetchError } from "./error";
 import {
   isPayloadMethod,
@@ -16,81 +8,13 @@ import {
   detectResponseType,
   mergeFetchOptions,
 } from "./utils";
-
-export interface CreateFetchOptions {
-  // eslint-disable-next-line no-use-before-define
-  defaults?: FetchOptions;
-  fetch?: Fetch;
-  Headers?: typeof Headers;
-  AbortController?: typeof AbortController;
-}
-
-export type FetchRequest = RequestInfo;
-export interface FetchResponse<T> extends Response {
-  _data?: T;
-}
-export interface SearchParameters {
-  [key: string]: any;
-}
-
-export interface FetchContext<T = any, R extends ResponseType = ResponseType> {
-  request: FetchRequest;
-  // eslint-disable-next-line no-use-before-define
-  options: FetchOptions<R>;
-  response?: FetchResponse<T>;
-  error?: Error;
-}
-
-export interface FetchOptions<R extends ResponseType = ResponseType>
-  extends Omit<RequestInit, "body"> {
-  baseURL?: string;
-  body?: RequestInit["body"] | Record<string, any>;
-  ignoreResponseError?: boolean;
-  params?: SearchParameters;
-  query?: SearchParameters;
-  parseResponse?: (responseText: string) => any;
-  responseType?: R;
-
-  /**
-   * @experimental Set to "half" to enable duplex streaming.
-   * Will be automatically set to "half" when using a ReadableStream as body.
-   * https://fetch.spec.whatwg.org/#enumdef-requestduplex
-   */
-  duplex?: "half" | undefined;
-
-  /** timeout in milliseconds */
-  timeout?: number;
-
-  retry?: number | false;
-  /** Delay between retries in milliseconds. */
-  retryDelay?: number;
-  /** Default is [408, 409, 425, 429, 500, 502, 503, 504] */
-  retryStatusCodes?: number[];
-
-  onRequest?(context: FetchContext): Promise<void> | void;
-  onRequestError?(
-    context: FetchContext & { error: Error }
-  ): Promise<void> | void;
-  onResponse?(
-    context: FetchContext & { response: FetchResponse<R> }
-  ): Promise<void> | void;
-  onResponseError?(
-    context: FetchContext & { response: FetchResponse<R> }
-  ): Promise<void> | void;
-}
-
-export interface $Fetch {
-  <T = any, R extends ResponseType = "json">(
-    request: FetchRequest,
-    options?: FetchOptions<R>
-  ): Promise<MappedType<R, T>>;
-  raw<T = any, R extends ResponseType = "json">(
-    request: FetchRequest,
-    options?: FetchOptions<R>
-  ): Promise<FetchResponse<MappedType<R, T>>>;
-  native: Fetch;
-  create(defaults: FetchOptions): $Fetch;
-}
+import type {
+  RequestInit,
+  CreateFetchOptions,
+  FetchResponse,
+  FetchContext,
+  $Fetch,
+} from "./types";
 
 // https://developer.mozilla.org/en-US/docs/Web/HTTP/Status
 const retryStatusCodes = new Set([
